Created Audio Manager
I have prefabed the Audio Manager, it's located in prefads with same name.

There is 2 new scripts one called sounds where the pitch, volume, loop, etc are kept and the Audio Manager which has the Array and Method for playing the sound.

if you need help with it msg me.